amazon-ads-api/CleanAndLoad/requirements.txt
boto3==1.34.75
bugsnag==4.6.2
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
pandas==1.5.2
python-dotenv==1.0.1
pyarrow==4.0.1
ijson==3.2.3
redis==5.0.4
simplejson==3.19.2
dask
psycopg2-binary==2.9.10
